Remove more Constants. (#4901)
PowChainService uses the genesis spec config for access to config as it needs to function prior to genesis and doesn't have a concept of slots.

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfig.java
@@ -128,7 +128,7 @@ public interface SpecConfig {
 
   int getMaxVoluntaryExits();
 
-  UInt64 getSecondsPerEth1Block();
+  int getSecondsPerEth1Block();
 
   int getSafeSlotsToUpdateJustified();
